requirements.txt
filelock==3.18.0
fsspec==2025.5.1
imutils==0.5.4
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
pandas==2.3.0
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
tzdata==2025.2
